+++ draft-packages/form/KaizenDraft/Form/Primitives/Radio/Radio.elm
@@ -1,4 +1,4 @@
-module KaizenDraft.Form.Primitives.RadioInput.RadioInput exposing (SelectedStatus(..), automationId, default, disabled, id, name, onChange, selectedStatus, value, view)
+module KaizenDraft.Form.Primitives.Radio.Radio exposing (SelectedStatus(..), automationId, default, disabled, id, name, onChange, selectedStatus, value, view)
 
 import CssModules exposing (css)
 import Html exposing (Html, div, input, text)
@@ -134,7 +134,7 @@ renderSelected status =
 
 
 styles =
-    css "@kaizen/draft-form/KaizenDraft/Form/Primitives/RadioInput/styles.scss"
+    css "@kaizen/draft-form/KaizenDraft/Form/Primitives/Radio/styles.scss"
         { radioInput = "radioInput"
         , box = "box"
         , icon = "icon"
