--- draft-packages/form/KaizenDraft/Form/Primitives/RadioInput/index.ts
@@ -1,1 +1,0 @@
-export * from "./RadioInput"
