+++ draft-packages/form/KaizenDraft/Form/Primitives/index.ts
@@ -5,4 +5,4 @@ export * from "./Input"
 export * from "./Label"
 export * from "./ToggleSwitch"
 export * from "./TextArea"
-export * from "./RadioInput"
+export * from "./Radio"
