--- draft-packages/form/KaizenDraft/Form/Radio/index.ts
@@ -1,1 +1,0 @@
-export { default as Radio } from "./Radio"
