+++ draft-packages/form/KaizenDraft/Form/RadioField/styles.scss
@@ -1,7 +1,7 @@
 @import "~@kaizen/component-library/styles/type";
 @import "~@kaizen/component-library/styles/color";
 @import "~@kaizen/deprecated-component-library-helpers/styles/type";
-@import "../Primitives/RadioInput/styles";
+@import "../Primitives/Radio/styles";
 
 .container {
   position: relative;
