+++ draft-packages/form/KaizenDraft/Form/index.ts
@@ -4,5 +4,5 @@ export * from "./Primitives"
 export * from "./TextField"
 export * from "./ToggleSwitchField"
 export * from "./TextAreaField"
-export * from "./Radio"
+export * from "./RadioField"
 export * from "./RadioGroup"
